Week_5_Warehousing_Streaming/openweather_basic_streaming/requirements.txt
pandas
kafka-python
google-cloud-bigquery
google-auth
pandas-gbq
pyarrow